Smaller vertex state for poplin ReduceAdd, use supervisor to pre-process it.
Summary: Reduction in size of vertex state, as per T4979.  This would have added 2 worker cycles to the execution time. BUT also moved the pre-processing of SCALEDPTR32 and division of work to the supervisor, passing it on the stack.  There is now a very small speed increase.

Test Plan: Re-run existing tests

Reviewers: samk, tomm, timh, samuelh

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D7358