Removed elementwise supervisor dispatch loop pipelining as it caused over reads well beyond the allowed limit
Summary:
Removed the loop pipelining in the case of supervisor dispatch loops, as the index
forward by (vector size * worker number) will cause overread beyond the allowed limit.  This would be the case for either the read before the loop (if no loop is expected) or for the pre-read within the loop (if the loop was taken).  For simplicity it was decided to just remove the pipelining.  Some speed penalty is expected.
Fixes T5415

Test Plan:
Re-run StdOperatorsTest
Re-run StdOpTest

Reviewers: samuelh, emmanuel

Reviewed By: emmanuel

Maniphest Tasks: T5415

Differential Revision: https://phabricator.sourcevertex.net/D7516