Adding broadcast operations: Add, Mpy, Subtract
Summary:
Added the optimisation whereby addInPlace, subInPlace and mulInPlace functions will detect if the 2nd operand is a scalar and the 1st is not. If so the scalar will not be broadcast by copying to memory, instead a "broadcast" version of the codelet will be used which uses the scalar tensor directly.  Supported by supervisor(1D) and worker (2D) codelets implemented in vectorised C similarly to the binary
and unary operations, with similar performance.  Added tests for the codelets and extended the
StdArithmeticTests to cover the new operations.  Added cycle estimates, each operation taking the
same amount of time - so using a simple single function for the estimates.
Fixes T4961

Test Plan:
Run StdArithmeticTests.cpp
Run broadcastOpTest.cpp

Reviewers: samuelh, samk, tomm

Reviewed By: tomm

Subscribers: jonc

Maniphest Tasks: T4961

Differential Revision: https://phabricator.sourcevertex.net/D7478