Scaled add with data, factor=half, delta=float mixed inputs
Summary:
Added a variant of scaled add which accepts data and factor with type=half, and deltas with type=float.
The emulate what would have happened previously with a cast of deltas to half before scaled add
was called, the codelet loads deltas, casts to half and then multiplies with factor and accumulates into data.
Data is then overwritten with the result.

Some optimisation of the supervisor assembly which divides the work.

The test data length was increased as the new variant of scaled add has an execution path that is only taken when
some workers have 4 x 4 halves to process.
Fixes T5012

Test Plan: Run ScaledAddSupervisor_fp and StdAritmeticTests

Reviewers: samk, tomm, samuelh

Reviewed By: tomm

Maniphest Tasks: T5012

Differential Revision: https://phabricator.sourcevertex.net/D7791