Added IpuModel tests for 2 ipus
Summary:
We (help/discussion with @tomm ) believe that the correct logic is used in getConvVertexTypeCandidates, which is allowed to select the outer product method, given that the transformed parameters have a batch size of 1. However the parameters which are passed to the cycle estimator don't include the correct transformation, so the check in addPartialCalcCycleEstimate erroneously concluded that batch size is not 1, producing the assert. As outer product should only be selected by the earlier code, it seems appropriate to simply remove the assert.

The same point holds that params.getNumInputChansPerConvGroup() is not transformed, whereas inChansPerGroup is.

This allows the selected tests to run without failure.
Fixes T5245

Reviewers: richard, tomm

Reviewed By: richard

Subscribers: tomm

Maniphest Tasks: T5245

Differential Revision: https://phabricator.sourcevertex.net/D8024