Reimplement pooling library in poplibs
Summary:
Supports general nD pooling operation.

The pooling library uses a planner which uses a grid
search to find parameters to split the pooling operation
in both forward and backward pass. The kernel is not
split over tiles in order not to perform a reduction
operation. The vertices used for avg/sum pooling for
both forward and backward pass are the same with an
additional selective scaling vertex for some cases of average
pooling.

The backward pass for an average/sum pooling layer which
have the same field and kernel size is detected and cast
as a broadcast operation.

To reduce exchange, a  second tile mapping scheme which
maps outputs to a tile based on regions of input tensor
used is supported. This is in addition to a linear tile mapping
scheme based on tile splits. Which one to use
is a pooling option.

The following is not implemented as part of this change
- nD support in pooling library in poplibs_support.
- Use of popsolver to perform planning
- Reductions after splitting(required if kernel were to be split)
- multi-ipu planning

The following result are produced for some benchmarks

** R18  B4 Train
|                      | Old               | New               | improvement
| -------------- | -------------- | --------------- |  -------------
| Total Alloc   | 232003304 | 221659071  | 10,344,233
| Vertex Data	| 63494645   | 52030688     | 11,463,957
| Exch              | 27107904   | 27943972     | -836,068
| Max Live       | 135631990 | 135895878  | 263,888
| Cycles           | 2350213      | 2367559      | -17,346

** R50 B1 Train

|                      | Old               | New            | improvement
| -------------- | -------------- | ------------- | -----------
| Total Alloc   | 204489591 | 202339621 | 2,149,970
| Vertex Data |40578051    | 37322875   | 3,255,176
| Exch	       | 34125184   | 34747536    | -622,352
| Max Live      | 110844920 | 111379188 | -534,268
| Cycles          | 2745092     |  2765218     | -20,126

** R50 B2 Train

|                      | Old               | New              | improvement
| -------------- | -------------- | -------------- | -----------
| Total Alloc   | 257679768 | 253418826 | 4,260,942
| Vertex Data | 49863221   | 44144508   | 5,718,713
| Exch             | 41524776   | 42143884   | -619,108
| Max Live      | 157714774 | 158122590 |-407,816

Fixes Maniphest: T2954, T5087

Test Plan: New tests added and all existing tests pass

Reviewers: godfrey.da.costa, Eddy, richard

Reviewed By: Eddy

Maniphest Tasks: T2954, T5087

Differential Revision: https://phabricator.sourcevertex.net/D7125