Reveal API for AVG, SUM pooling that only requires one tensor - redone following pooling rewrite
Summary:
Added separate API, however the input Tensor was used to determine the dimensions and mapping of the result tensor. Therefore dimensions of the desired result tensor are taken from the poolParams structure and passed as required in the separate API.  This is a re-work of D7847 as that diff clashed with a rewrite of a lot of the code that it affected, and was reverted.
Fixes T5079

Reviewers: samk, tomm

Reviewed By: samk

Maniphest Tasks: T5079

Differential Revision: https://phabricator.sourcevertex.net/D8229