Handle forward averaage/sum pooling as a reduction where possible
Summary:
When the kernel and field shapes match, sum/average pooling
may be represented as a reduction operation. This results in
a slight increase in memory but is a lot more efficient in cycles
than  the current average pooling assembler impementation. With
the current assembler implementation, for small batch sizes,
the amount of tiles over which the last RESNET layer can be
spread is not optimal because of the constraint on the grain
size and the fact that kernel splitting is not enabled.

Improves RESNET18 batch size 1 performance inference by ~3.4%

T5981

Test Plan: ./test.sh poplar

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D8305