Add support for Group, Instance and Layer normalisations in popnn
Summary:
The code is restructured to move batch norm out of
the convolution.cpp file. Support for fully connected
and convolutional layers is unified with support for nD
convolutions. Add channel ops is still part of poplin,
the normalisation functions are still part of the library.
These can be removed once T4987 is addressed.

T6087

Test Plan: New tests added for Group, Instance and Layer normalisation

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D8249