Handle cases when number of input channels to convolution vertices is zero
Summary:
If the number of input channels is zero, there could
still be zero output produced. The 1x1 vertex doesn't
produce output if the number of input channels is 0.
In this case, use an nx1 vertex. The "numInChanGroups"
vertex field now gives the exact value rather than 1
less. This could have been handled by setting
inChansPerGroup to be zero when number of input
groups is zero; which could have saved a few cycles
in the codelets. This is not done to keep the meaning
of the fields clear but could be changed to save a
minuscule number of cycles/memory.

T5824

Test Plan: ./test.sh poplibs including  Sim tests which time out

Reviewers: Richardc

Reviewed By: Richardc

Differential Revision: https://phabricator.sourcevertex.net/D8795