Improve tile balance for norms using dithering on statistics and parameters mapping
Summary:
In norms the statistics and parameters tend to get mapped to the same
set of tiles. With this change, a dithering is applied on the start
tile used for mapping statistics and parameters. The dithering is
randomised based on the gradient and activations dimensions and
would give the same mapping for the same shape of gradient and
activations. This implementation is not yet optimised for multi-IPU
where there is a potential benefit of introspection.

R50, GN with num groups=32
BS1:
    Old
        Mean tile memory: 211555
        Max tile memory (tile 2): 253462 (119% of mean)
    New
        Mean tile memory: 211518
        Max tile memory (tile 50): 240948 (113% of mean)
BS2:
    Old
        Mean tile memory: 260661
        Max tile memory (tile 5): 295899 (113% of mean)
New
        Mean tile memory: 260657
        Max tile memory (tile 314): 287172 (110% of mean)

Test Plan: ./test.sh poplibs

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D8875