Scale partials in 32-bit precision and then reduce precision
Summary:
Reductions in popops should scale partials and then
reduce precision rather than the current method
of reducing precision and then scaling. This change
applies this for FLOAT to HALF reductions only and
not for HALF to HALF as there is a cost involved.

T6809

Test Plan: ./test.sh poplibs

Reviewers: samk, O1 Reduction

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D9295