Take preprocessed tensor groupings from the plan instead of introspection. Allow WU rearrangement to catch resnet 7x7 bs1 layers.
Summary:
Fixes T6684.
As titled. The desired grouping is actually literally just the input channel grain size as things stand due to the further transforms that exist and the way that we map regrouping operations.
This brings us almost all the way back down to where performance was prior to WU rearrangement changes.

Fixes T6430.
We also allow the WU rearrangement to add padding in a way that results in it being interleaved in the resulting input channels due to further dimension expansions. This only really affects resnet 50 batch size 1 7x7 layers, makes a minor difference.

Reviewers: godfrey.da.costa, richard

Reviewed By: godfrey.da.costa

Maniphest Tasks: T6684, T6430

Differential Revision: https://phabricator.sourcevertex.net/D9243