Scatter - Support different update methods
Summary:
Added support for an update computation when using scatter.
This is useful for combining the existing values in the input array and the updates during a scatter, which is required in TF and ONNX gather gradient.

Resolves T6134.

Test Plan: Existing scatter tests pass and added a new test case with an update computation.

Reviewers: #software_-_poplibs, richard

Reviewed By: #software_-_poplibs, richard

Subscribers: richard, davidn, georgep

Maniphest Tasks: T6134

Differential Revision: https://phabricator.sourcevertex.net/D9571