Add support for scaled gradient in MAX pool backward pass
Summary:
Scaled gradient is a better mathematical representation
of the max pool gradient. This change implements this
as an option.

A scale tensor is computed in the backward pass even
though it could have been in the forward pass cheaply.
But because of the need to hang on to it for the backward
pass, the scale is computed in the backward pass. The
scaling itself is done using an element wise op, but it
may be better to fuse it with the gradient calculation
if the assembler implementation is cheaper (T7133).

T6960

Test Plan: Added new tests for max pool with scaled gradient

Reviewers: samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D9419