Make tensor regrouping function public in poplin interface and use it in Norms
Summary:
Operations on two tensors with the same grouping is
better suited for certain operations. Typically convolution/matmul
outputs in the fwd and bwd pass may be regrouped in differently.
A regrouping function introspects and checks for grouping on
two tensors - an input and reference tensor - and if they are not
grouped in the same dimensions rgroups the input tensor.

This is currently applied to norms but may be extended to
other ops which typically use convolution input/output. The
regrouping function may be further generalised to be part of
a rearrangement library but is not done in this change.

Test Plan:
./test.sh poplibs tests pass

Partially addresses T7132

Reviewers: tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D9811