Add template parameter to reduction vertices to indicate single element regions
Summary:
The reduction output edge alignment is always set to 8. This allow
usage of vector (v4/v8 for half and v2 for float) instructions.
However there are several cases where the output regions may have a
single  element. In this case, although the code exists in the
assembler to handle it, it is still aligned at 8 just because
the edge definition in the vertex is so.

In several models, there may be only one output element per
output region. Using the existing vertex as such would mean
there will be post arrangement copies just to satisfy the alignment
constraint.

A template parameter to the reduction vertices to indicate single
element regions. This can be set at graph construction time based
on the detection of single element regions. The same assembler code is
used by calling a common function.

T7372

Test Plan: all tests pass

Reviewers: samk, O1 Reduction

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D9877