Guarantee that 1x1 convolution vertices have at most a single partition
Summary:
The 1x1 convolution for most models has a single partition
per worker. Having a general function increases the overhead
of dealing with partition pointers. Instead of specialising the
case with a codelet, this change makes 1x1 vertex only deal
with a single partition per worker. If there are more than
one partitions, an nx1 vertex is selected.

Saves both memory and cycles for RESNET with small batch
sizes.

Test Plan:
./test.sh poplibs with 5000 tests (there are some
test failures in Sim but all these are because of OOM.

Reviewers: Eddy

Reviewed By: Eddy

Tags: #software_-_poplibs

Differential Revision: https://phabricator.sourcevertex.net/D10344