Improve add, subtract and mpy elementwise ops where broadcasting is required, including move channelMul, addToChannel to popops
Summary:
This is an optimisation of the case where we are doing a binary add, multiply or subtract operation, when the 2nd operand is a vector like tensor (only one dimension is not of size 1).  The approach is to make an estimate of the cycle count for implementing the operation using a variety of methods, including channelMul, addToChannel, the scalar operation broadcast cpp codelets, binary op codelets and the previously used "copy to broadcast" method.   The method that is available with the least cost is the one that is used.

addToChannel, channelMul codelets are now in popops, as they needed to be called from popops.

Using add, sub, mul, mulnPlace, addInPlace and subInPlace will now choose the best method available. However the requirement for in place/non-in place operations constrains the choices.

Performance improvements were measured with the BroadcastOptimiseTest tool, and also in the context of resnet. Testing with resnet was focussed on making improvements to the case where  D8432 was removed from the diff and the following b renchmark run:

enigma/benchmarks/resnet --variant RESNET_50 --mode TRAIN --norm-type GroupNorm --num-norm-groups 32

This resulted in a size increase/performance decrease.  Changing relevant references to channelMul, addToChannel to mul, addInPlace etc results in optimised choices being made, examples of the resulting size decrease/ performance increase are included below.

Test Plan: New test tool BroadcastOptimiseTest, with a variety of test options which were observed to exercise generation of every codelet type.

Reviewers: tomm

Reviewed By: tomm

Subscribers: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D9399