Plan implementation of element-wise ops on a per-tile basis.
Summary:
Fixes T7848.

This is to allow use of broadcast vertices on each tile.
This should be an improvement in cycles if not necessarily memory usage for element-wise ops with broadcasted operands due to the fact that these would currently be implemented with a copy and then the element-wise operator.
The copy is very slow as it broadcasts one atom at a time to the destination through a strided copy vertex.

Currently this is very simple, where limitations are currently I have left comments. The framework for whatever different types of broadcasting patterns you might want to specialise for should all be in place.

This change also includes a bunch of tidy-up for broadcast codelets and element-wise library code.

Test Plan: Let CI run the lot.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Subscribers: claudiop, timp

Maniphest Tasks: T7848

Differential Revision: https://phabricator.sourcevertex.net/D10705