Add new random library interface which allows seed setting as an option
Summary:
The existing interface required the user to set the seed
before any operation that required a fixed sequence to be
generated. The new library allows the user to either set
seed or not. If the seed is set, the h/w random generator
state is restored after the random operation is performed.

Clean-up of tests, bug fixes etc.

T7588, T7546

Test Plan: Added new tests

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: davidn, badreddine

Differential Revision: https://phabricator.sourcevertex.net/D10447