Add a TransposeSupervisor<half> vertex
Summary:
A couple of functions in Convolution.cpp and ConvUtil.cpp perform the transposition of many small matrices (on multiple tiles)
On each tile multiple worker are created to do the work.

This change will add a TransposeSupervisor vertex that can be called once on the tile and it will provide to partition the work best among workers.
This will add some extra cycles (for the supervisor) but save space in memory (as there will be a single vertex state per tile instead of one per worker).

Reviewers: timp, samuelh

Reviewed By: timp, samuelh

Subscribers: timp, samuelh

Maniphest Tasks: T6398

Differential Revision: https://phabricator.sourcevertex.net/D10442