Simple optimisations to scaled reduction code, and add scaling option to specialisation 3
Summary:
Noted that more cases could be targeted by specialisation 3 (which is fast) if it supported scaling.
Also noted inefficient load of the scaling.

Running `resnet --mode TRAIN --batch-size 4 --norm-type BatchNorm` shows that instances of ScaledReduce specialisation 0 become specialisation 3.  Around 0.3 MByte saving, and 27 uS faster. (Note that this is a big piece of what is saved in D8170.  The changes in that are ongoing, however it seemed good to land this easy saving before they are completed.)

Reviewers: O1 Reduction, samk

Reviewed By: samk

Differential Revision: https://phabricator.sourcevertex.net/D10914