Align input to OuterVector non-inplace version.
Summary:
After D10833 there are non-inplace broadcast codelets, the inner kernel for the BroadcastOuterVector vertex does not handle differing alignments between the input and output pointers.

This change just makes BroadcastVectorOuter have an 8-byte aligned input so it will always have the same alignment as the output pointer in the inner kernel.

Also switch inOutType/dType in vertex definitions as it seems like this was the intended way round.

Reviewers: claudiop, timp

Reviewed By: claudiop, timp

Differential Revision: https://phabricator.sourcevertex.net/D11203