Increase the number of cases in which scalar broadcast operations can be used
Summary:
Begin by splitting up the patterns that are detected on each tile into "single region" patterns - those that can be created using a scalar broadcast vertex.  There can in some cases be a vast number of these, so abort if that is the case. Where there are a reasonable number of patterns, and each worker has a reasonable amount of work, then use the scalar broadcast ops to implement the operation.  Otherwise exit, allowing other methods of carrying out the operation to take over.

Fixes T8486

Reviewers: tomm, samuelh

Reviewed By: tomm

Maniphest Tasks: T8486

Differential Revision: https://phabricator.sourcevertex.net/D11209