Added the ability to enable shared structures for the single_conv_layer tool
Summary:
Also added the option to the random test generator so it gets tested as part of a normal `./test.sh poplibs` run.

Also enabled the multi_ipu variant for the Sim variant.

Ref T6583. Depends on D10541, D11145, D11172.

This change uncovers a bug and so is blocked by T8396 before it can be landed.

Test Plan: `./test.sh poplibs` with a temporarily increased amount of conv random tests. Also observed manually the options were used as expected.

Reviewers: tomm, Eddy, godfrey.da.costa, richard

Reviewed By: richard

Maniphest Tasks: T6583

Differential Revision: https://phabricator.sourcevertex.net/D10835