Consider both joint and separate fully connected plans and pick the best
Summary:
Previously if the pass was set to Pass::FC_TRAINING_* the planner always
chose a single plan for the forward / backward / weight update passes that
was guaranteed not to exchange of weights. This is a good strategy if the
weight matrix is large and amount of compute is small as in this case
exchanging weights is very costly.

If the amount of compute is large and the weight matrix is small a joint
plan could be inefficient as it restricts the freedom of the planner to
select the best method for each pass. Update the planner so it considers
both joint plans and separate plans picks whichever has the lowest cost.

Fixes T8346

Reviewers: Eddy, godfrey.da.costa

Reviewed By: Eddy

Maniphest Tasks: T8346

Differential Revision: https://phabricator.sourcevertex.net/D11153