WriteUndef the full tensor which constituent parts are written by the serialised convolution
Summary:
The serialisation of a big convolution to save temporary
memory uses a dynamic update of the partial results into
the full tensor. This leaves the full tensor live
throughout the program. WriteUndef the full tensor to
prevent liveness analysis from keeping it live.

Test Plan: CI. Checked that the full tensor is not kept live.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D11305