Clone input and output LSTM weights together
Summary:
If options.preCalcWeights is false the input weights and output
weights are slices of the same variable. Update createWeightAccumulators()
to clone them together to ensure the input and output weight gradients are
also slices of the same variable. This simplifies tensor expressions
involving the concatenation of the input and output weight accumulators.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Subscribers: jakeh

Differential Revision: https://phabricator.sourcevertex.net/D11476