Improve reduction memory usage and cycles by supporting different data patterns with different vertices.  Resulting in a new vertex type, "PartialsEqualSize".
Summary:
A specialisation of the reduction vertices which targets the case where:
The output width is a multiple of the grainSize for the input type (8 halves, 4 floats)
There is a single output region
All partials have the same width which is a multiple of the output width.
This results in the name "PartialsEqualSize".

As the typical cases observed have a small output width, and partial width (Eg 8 halves - see comment from Jan 7) the optimisation is to sum down the 'column' of the partials.

Some background - initially this was started some time ago.  The Jan 7th comments by samk illustrate the sizes of the reductions required following some investigation.  These were either:
1. Wide inputs - a small number partials, each with many elements
2. Narrow inputs - a large number of partials, each with very few elements

In both cases 1 and 2 the widths were nice multiples of 8 and so was the output.  Two types of reduction vertex were proposed - Row major order (RMO) for case 1 and Column major order (CMO) for case 2.  At this point it was shelved.  In the meantime Eddy produced various specialisations of the reduction vertices based on what he saw.  Specialisation 3 serves the same purpose as RMO.  That left this diff to produce CMO.  However the naming convention now makes less sense.  The target of CMO is the "2. Narrow input case", so the effect of adding the CMO option is that data does not need to be gathered, which was the case when it was targeted by specialisation 3.  CMO is however somewhat slower due to having to unpack pointers in the inner loop.  The tradeoff should hopefully be dealt with in vertex generation.

Eventually CMO was renamed PartialsEqualSize.

Part of T2860.

Test Plan: `./test.sh poplibs -R reduce`

Reviewers: O1 Reduction, godfrey.da.costa, samuelh, Eddy, samk

Reviewed By: samuelh, samk

Subscribers: samk, O1 Reduction, timp, samuelh, godfrey.da.costa

Maniphest Tasks: T2860

Differential Revision: https://phabricator.sourcevertex.net/D8170