Create and select always aligned/misaligned versions of VectorOuter broadcasting supervisor vertices.  Bug fix to supervisor dispatch.
Summary: As per title.  However a difficulty was discovered in the case of vertices dealing with misaligned data.  Data in one contiguous region is processed, row by row.  Where the number of halves in a row is odd, one worker can process the end of a row, and another worker the beginning of the next.  The read modify write on the overlapping 32 bits can cause a conflict. This is a potential problem in both the case where work is split by assigning all workers to a row at once or when assigning each worker a single row.  The case of all workers assigned to a row at once was an existing case which looked wrong but hadn't (yet?) caused a known issue.  This is the bug fix to the supervisor dispatch.  The case where each worker is assigned a row is dealt with by assigning consecutive pairs of rows to each worker.

Reviewers: tomm, claudiop, samuelh

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D11358