Create inplace/non-inplace variants for all Channel Ops, and rename for consistency
Summary:
The 4 'Add To Channel' codelets, those that do broadcast operations on the
innermost (rightmost) dimension, i.e.:
  'AddToChannel',       'AddToChannel2D'
  'ScaledAddToChannel', 'ScaledAddToChannel2D'
are all in-place (i.e. the output goes into the first operand).
We want to have non-inplace variants for those, and to target them when
possible.
Similarly the 2 'Channel Mul' codelets ('ChannelMul', 'ChannelMul2D') are
non-inplace only, so we might want in-place variants for them.

Also, we want to rename all of them following thr pattern of the other
broadcast codelets:
  BroadcastVectorInner[Supervisor/2D][InPlace]

Reviewers: timp

Reviewed By: timp

Maniphest Tasks: T8694

Differential Revision: https://phabricator.sourcevertex.net/D11556