Variable learning rate applied to batchNormParamUpdate
Summary:
Implemented for all xxxNormParamUpdate functions - float and Tensor learning rate supported, renamed scale.

Note that scale (as a tensor) needs to be initialised with -learningRate.  Therefore for consistency (and common with other functions which accept either a constant or variable learning rate) the constant scale (float) now needs to be passed as -learningRate, not learningRate as before.

I can't see evidence that these functions are actually used, and apparently are not used in tensorflow/onnx, so maybe they could be removed, but they are very simple, and easy enough to leave in the codebase.

Fixes T6512

Reviewers: claudiop, samk

Reviewed By: samk

Maniphest Tasks: T6512

Differential Revision: https://phabricator.sourcevertex.net/D11785