Add implementation of all-reduce that uses replicated copies.
Summary:
The new implementation keeps the tensor as a replicated tensors and uses
cross replica copies to copy between replicated instances. By default the
old implementation is still used - the new implementation can be enabled by
specifying the useReplicatedImplementation option.

The new implementation limits the use of getNonReplicatedTensor to
the replicatedRankSlice() and replicatedRankUpdate() functions. In
future the implementation of these functions will be changed so the
slice being accessed is determined at runtime based on the value of
a tensor that is initialized to the rank. At that point the code
for each replica will be identical which lets us compile one image
that can be loaded multiple times.

The main differences compared to the old implementation are as follows:

- We add padding to ensure all fragments are the same size.
- The same buffers are used at each step of the collective which lets
  us reuse the code for copying between IPUs / performing the reduction.
- There is some additional copying, in particular the first reduce scatter
  fragment is copied to the srcBuffer before it is sent to the neighbouring
  IPU and the all gather fragments are written to the dstBuffer before
  they are copied to the final location.

I expect this to be slightly slower than before (due to the extra copies)
and more memory efficient (due to better code reuse).

Ref T7820

Reviewers: samuelh

Reviewed By: samuelh

Maniphest Tasks: T7820

Differential Revision: https://phabricator.sourcevertex.net/D11700