Port model pooling to take vectors of dimensions
Summary:
Ref T5261.

Currently the dimensions of stride, padding, etc.. is hardcoded to 2. This diff allows an arbitrary amount of dimensions to be pooled, although note that it does not introduce any pooling tests that use more dimensions; that will arrive in a separate diff.

To support the existing use-cases of the multi-array data in the model pooling implementation we needed to support copying a slice of the multi-array. This means that the assignment operator for a `{Const}Reference` needs to handle two cases at runtime:
 - Is the reference we're assigning from a partial slice (ie it's shape is not empty)? Copy the reference.
 - Otherwise (the reference is pointing at a single element) copy the value.
An exception is thrown if we try to assign a reference pointing to an element to a reference of a partial slice. This matches the behaviour of `boost::multi_array` except that was able to verify this at compile time -- something we can no longer do as the rank is not known until runtime.

Another note, this solution is not as efficient as it could be as we have to build up a new shape (a vector) for each permutation of a multi-array's indices if we want to read from another multi-array but transforming the indices which we do a lot. We can improve the effeciency by supporting a tranform function that can change an index based on another without storing the whole set. The existing pooling tests didn't seem to suffer from this too much so it was not done as part of this diff.

Test Plan: `./test.sh poplibs -R pool`

Reviewers: godfrey.da.costa, timh, timp, claudiop, Eddy, richard

Reviewed By: richard

Maniphest Tasks: T5261

Differential Revision: https://phabricator.sourcevertex.net/D11600