Add the ability to partially constrain a convolutional plan through option flags.
Summary:
Fixes T9402.

You can now pass a JSON string as the option "planConstraints" for both MatMul and Convolution libraries.
If a valid plan with the given constraints can't be found, you will get an error, this will not fallback.

The form of the plan constraints is e.g.:
```
{"0", {"transform": {"swapOperands": true},
       "partition": {"fieldSplit":{"1": 4},
                     "inChanSplit": 4,
                     "outChanSplit": {"parallel": 4}}
      }
}
```
Where the outer-most index in the plan is an index into the plan hierarchy, and any multi-dimensional fields are sparsely indexed objects i.e. constraining dimension 1 of fieldSplit to be 4 is specified as `{"fieldSplit": {"1": 4}}`.

This is only implemented for partitioning and for the swapOperands transform for now. This is just to save time implementing everything up front, adding more should be easy.

Also sync up OptionParsing code with poplar.

Reviewers: richard, samk

Reviewed By: samk

Maniphest Tasks: T9402

Differential Revision: https://phabricator.sourcevertex.net/D12134