Add iota function to popops
Summary:
The vertices only have C++ codelets as the compiler is expected
to do a decent job in code generation. An optimisation which is
not done is to introspect and use a constant tensor copy if the
number of elements on tile are below a certain limit.

T9453

Test Plan: Added new tests

Reviewers: tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D12220