Speed up LSTM zeroWeightAccumulators
Summary:
popops::zero() calls reorderToSimplify() on the inputWeights and outputWeights. Since these are slices of dimshuffles, it can't simplify the expression successfully and produces an expression where every single element is a complex slice. This change recombines the slices so that they can be simplifies successfully. It reduces the runtime for `encoder_decoder --mode TRAIN --device-type Sim --batch-size 4` from 4:05 to 1:38 (which is still a bit slow, but much better!).

Fixes T9596

Test Plan: CI & timing encoder_decoder

Reviewers: georgep

Reviewed By: georgep

Maniphest Tasks: T9596

Differential Revision: https://phabricator.sourcevertex.net/D12325