Update collective code to work with single image replication
Summary:
Call Graph::getTopLevelGraph() to determine whether the top level graph
is replicated or not. If the top level graph is replicated use a switch
to select which copy to perform, otherwise use getNonReplicatedTensor()
to build a copy that accesses the correct slice in each replica.

This change was originally committed in cd8d1fdbf428 but it was reverted
because it exposed a bug in getTopLevelGraph()

Depends on D12298
Fixes T7820

Reviewers: samuelh

Reviewed By: samuelh

Maniphest Tasks: T7820

Differential Revision: https://phabricator.sourcevertex.net/D12326