Zero out gradients for masked label elements
Summary:
The gradients for masked labels must be set to zero.
The implementation for doing this is dicrete using
comparison, cast and broadcast tensor multiplies.
If seen as a performance/memory issue, the on-tile
transforms may be modified to pass up the label per
batch.

Test Plan: Modified test to guarantee masked label generation

Reviewers: tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D12337