MSR Add SelectScalarFromRows and UpdateScalarInRows
Summary:
Add implementation of two new poplib function that provide the building blocks of the TF op tf.nn.sparse_softmax_cross_entropy_with_logits
https://www.tensorflow.org/api_docs/python/tf/nn/sparse_softmax_cross_entropy_with_logits

SelectScalarFromRows:
For each row in the 2D tensor params, select a single scalar value.
Aggregate the resulting scalars into a 1D tensor.

UpdateScalarInRows:
Update in-place one scalar per row of the tensor params. For each row,
the index of the value to update is specified by the tensor indices.

The implementation of UpdateScalarInRows is fine-tuned to the specific needs of the sparse cross entropy computation.
It can be extended in the future when the need arises.

Thanks to Juliana, Tiyasa and Renato for discussing this with me.
T8773

Reviewers: jakeh, tomm

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D12035