Prevent createPartialsLevel from going above a serial split.
Summary:
Ref T9401. Depends on D12762.

This is a compile-time optimisation which aims to reduce the number of individual concatenated variables in a tensor expression by creating one larger variable for partials at a higher level. When we have a serial split we will be updating the result of this sub-convolution that is split in parallel (the partials/output from this serial slice) into another larger output. The main optimisation is to create one larger set of partials for each individual tile in a parallel convolution, and as such this seems okay for the time-being as the split between IPUs higher up introduces fewer partial variables.

Reviewers: richard, samk

Reviewed By: samk

Maniphest Tasks: T9401

Differential Revision: https://phabricator.sourcevertex.net/D12771