Always copy gathered data from the same buffer, reducing code size.
Summary:
The all-gather code uses two buffers. Previously the role of the two
buffer alternated in each step - the buffer that was the destination
of the previous step became the source in the next step. When single
image compilation is enabled we included code to copy from each buffer to
every fragment of the result, even though a particular replica would only
execute half these copies.

After this change we always use the same buffers to send and receive and
so we only ever copy from the one of the buffers to the result, halving
the amount of copy code required. The downside is we need an extra copy
from the destination buffer to the source buffer between each
step, slightly increasing runtime.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D12548